+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Improve predicate call parsing performance when using the code navigation features
 * Fix error when calling the "Load File" and "Load Directory" commands from the command palette when no file is open or selected
 * Fix potential race condition when loading a project on activation
+* Fix displaying of inline code coverage data on Windows
 
 ## [0.79.0]